GUI-2120 Create dialog.
